requirements/dev.txt
-r base.txt
ipython==8.25.0
ipdb==0.13.13
psycopg==3.2.3
psycopg-binary==3.2.3
djangorestframework==3.15.2
django-extensions==3.2.3
requests==2.32.3
tzdata==2024.1
pillow==10.4.0
reportlab==4.2.5